+++ Recon/include/Recon/Skims/RecoilFiducialityProcessor.h
@@ -9,7 +9,7 @@
 
 // LDMX
 #include "Ecal/Event/EcalHit.h"
-#include "Event/FiducialFlag.h"
+#include "Recon/Event/FiducialFlag.h"
 #include "Framework/Configure/Parameters.h"
 #include "Framework/EventProcessor.h"
 #include "SimCore/Event/SimCalorimeterHit.h"
